Puppet: Move parts and related types/functions into their own separate file
